+++ tests/test_settings/window.h
@@ -1,0 +1,23 @@
+#ifndef TESTWINDOW_H
+#define TESTWINDOW_H
+
+#include <QMainWindow>
+#include <QWidget>
+#include <QVBoxLayout>
+#include <QCheckBox>
+#include <QLineEdit>
+
+class MainWindow : public QMainWindow
+{
+    Q_OBJECT
+
+public:
+    MainWindow(QWidget *parent = nullptr);
+    ~MainWindow() override;
+
+public:
+    QLineEdit *ptrLineEdit;
+    QCheckBox *ptrCheckBox;
+};
+
+#endif // TESTWINDOW_H
